chapter_02/requirements.txt
openai==2.30.0
python-dotenv==1.2.1